tests/wpt/tests/svg/scripted/script-invalid-script-type.html

<!doctype html>
<link rel="author" href="mailto:avandolder@mozilla.com" title="Adam Vandolder">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1827512">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="host"></div>
<svg><script type="text/plain">window.scriptRan = true;</script></svg>
<script>
  let svgScript = document.querySelector("svg > script");
  svgScript.type = "text/javascript";
  svgScript.appendChild(document.createTextNode(""));
  test(function() {
    assert_true(window.scriptRan);
  }, "svg:script runs after valid script type is set");
</script>
